### 2020-09-02 #2634 @whedon
[PRE REVIEW]: Empirical and non-parametric copula models with the `cort` R package
**Submitting author:** @lrnv (<a href="http://orcid.org/0000-0002-7508-999X">Oskar Laverny</a>)
**Repository:** <a href="https://github.com/lrnv/cort" target ="_blank">https://github.com/lrnv/cort</a>
**Version:** 0.3.1
