### 2020-09-02 #2634 @kthyng
Hi @lrnv, thanks for your submission. Upon looking through your paper, I see that it does not [have](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)

> A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience.

I cannot as a non-expert in your field understand enough to pass this onto a handling editor. Please make this change and ping back here when that is finished. While you are at it, please add the missing doi's listed above, too.
