+++ paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={525--531},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ejor.2015.06.028}
 }
 
 @inproceedings{ram2011density,
@@ -25,7 +26,8 @@
   number={4},
   pages={877--889},
   year={2018},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/03610926.2017.1285929}
 }
 
 @article{sklar1959fonctions,
@@ -44,7 +46,8 @@
   volume={151},
   pages={69--89},
   year={2016},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jmva.2016.07.003}
 }
 
 @article{li2018panda,
