### 2020-09-03 #2634 @lrnv
I did correct the DOIs and rework the paper. Tell me if you think it's still not clear -- As the software solves a research-level problem, it's quite hard to describe to non-specialist, but i tried.
