+++ paper.md
@@ -20,9 +20,11 @@ bibliography: paper.bib
 
 # Summary
 
-The R package `cort` implements object oriented classes and methods to deal with specific statistical objects: copulas. Copulas are distributions functions on the unit hypercube that have uniform margins (what we call the 'copula constraints'), and hence this package can be classified in 'density estimation software'.
+The R package `cort` implements object oriented classes and methods to estimate, simulate and visialise certain types of non-parametric copulas.
 
-Although the estimation of copulas is a wide-treated subject, most performing estimators available in the literature are based on restricted, parametric estimation: vine copulas [@nagler2016evading] and graphical models [@li2018panda] for example are potential solutions but under restrictive assumptions. Classical density estimators such as kernels or wavelets do not satisfy marginal copula constraints. There also exists several tree-structured piecewise constant density estimators, but they do not always lead to proper copulas when applied on pseudo-observations or true copula samples.
+Copulas are functions that describe dependence struture of a given dataset, or of a given multivariate random event, without describing the univariate events, the marginals. In statistics, It is sometimes usefull to separate the marginal distributions (inflation of the money, mortality of the population) from the dependence structure between them, sine estimating everything separately is usualy easier. Copulas are broadly used in finance, actuarial science, geostatistics, bistatistics, and many other field, when dealing with dependence.
+
+Copulas are distributions functions on the unit hypercube that have uniform margins (what we call the 'copula constraints'), and hence this package can be classified in 'density estimation software'. Although the estimation of copulas is a wide-treated subject, most performing estimators available in the literature are based on restricted, parametric estimation: vine copulas [@nagler2016evading] and graphical models [@li2018panda] for example are potential solutions but under restrictive assumptions. Classical density estimators such as kernels or wavelets do not satisfy marginal copula constraints. There also exists several tree-structured piecewise constant density estimators, but they do not always lead to proper copulas when applied on pseudo-observations or true copula samples. The new models that are implmeented in this package tries to solve these issues.
 
 
 # Statement of need 
