+++ pkgdown.yml
@@ -6,5 +6,5 @@ articles:
   vignette02_cort_clayton: vignette02_cort_clayton.html
   vignette03_ecbkm: vignette03_ecbkm.html
   vignette04_bootstrap_varying_m: vignette04_bootstrap_varying_m.html
-last_built: 2020-09-03T15:29Z
+last_built: 2020-09-03T15:30Z
 
