### 2020-09-03 #2634 @lrnv
@kthyng What do you think about this last version ? If this is still not enough, i'll rethink it dont hesitate to tell me.

Thanks for taking care of that :)
