### 2020-09-07 #2634 @pdebuyl
@coatless @emilyriederer @agisga would you review the paper "Empirical and non-parametric copula models with the `cort` R package" for the Journal of Open Source Software?
