+++ pkgdown.yml
@@ -1,10 +1,6 @@
 pandoc: 2.7.3
-pkgdown: 1.5.1
+pkgdown: 1.6.0
 pkgdown_sha: ~
-articles:
-  vignette01_ecb: vignette01_ecb.html
-  vignette02_cort_clayton: vignette02_cort_clayton.html
-  vignette03_ecbkm: vignette03_ecbkm.html
-  vignette04_bootstrap_varying_m: vignette04_bootstrap_varying_m.html
-last_built: 2020-09-07T04:30Z
+articles: []
+last_built: 2020-09-08T04:30Z
 
