### 2020-09-09 #2634 @pdebuyl
@agisga the expected review time is currently of six weeks. It is usually a bit shorter (four weeks) but is extended until the working conditions improve in regard to the covid-19 pandemic. Let me know if that would work for you.
