### 2020-09-09 #2634 @coatless
@pdebuyl more than happy to review the software! I'm excited to see an improvement in copula estimation within _R_.
