### 2020-10-12 #2653 @pdebuyl
@coatless @agisga a complete review is expected by the middle of next week. Intermediate questions and comments to the author are also welcome on this page, as the review process of JOSS enables this type of communication.
