### 2020-10-22 #2653 @pdebuyl
Thank you for the review @agisga !

@lrnv if you would like to see examples of community guidelines, for instance, you can take a look at already published JOSS articles.
