### 2020-10-22 #2653 @lrnv
Well, first, thanks to you two @coatless and @agisga for taking the time. At first glance, the review you produced are very interesting, and i can only nod to every point you highlighted.

The plan i have right now:
1° Fix issues that the reviewers highlighted (and more if i find more) in a new branch on my repo, and prepare a Pull request. This will take me a few weeks.
2° Wait for the reviewers to see the Pull request and eventualy discuss it on it's own thread.
3° Once everyone agree, merge the PR into master and send the package to CRAN
4° Once CRAN agree (again, a few weeks delai), accept here the new version of the paper (with the right verison number and everything..)
5° Continue the publication in JOSS after that.

@pdebuyl do you agree with this plan ? Please tell me what you think.
