### 2020-10-22 #2653 @pdebuyl
@lrnv Your plan is fine. You will need to archive the code to zenodo as well, though. This is part of the JOSS process and the zenodo archive DOI will be included in the article. I will also perform a final proofreading before forwarding the article to an editor in chief for publication. The paper might thus cause a few more commits to the git repository but this won't change the software itself so that the CRAN and zenodo version are good.

Typo, page 2, first line. "efficicent" -> "efficient"

Typo, page 2, acknowledgments. "meaningfull" -> "meaningful".

Grammar, page 2, second paragraph. "by statistician" -> "by statisticians" (if you pick singular instead, then conjugation of "need" should be changed as well).

Affiliation: please add "France" to the first affiliation and spell out the second affiliation (+ add city and country).

For specific issues such as the API naming, you can discuss them in cort issues and report the conclusion here. For changes to the paper, I prefer if they are done in this thread.
