### 2020-11-03 #2653 @pdebuyl
Hi @lrnv we have no strict timeframe once the reviews are in, as it is often harder to get the reviews than the author's update. Still, to avoid that the reviewers have forgotten about the package when you reply, it is best to avoid unnecessary delays. Let us know when there is an update.
