### 2020-11-12 lrnv/cort#6 @lrnv
@coatless review
The review from @coatless:

I'm a huge fan of the package. Long ago I stepped foot into copula estimation with _R_ and was disappointed by the tools available. `cort` fills portions of that void. Examples with the package are easily accessible outside of _R_ thanks to a `pkgdown` website. Having said this, I do have a few comments and suggestions around the package. I've listed them below.

### README

#### Installation

Please indicate on the README that installation from GitHub will require the system to have a compiler.

- Windows:  Rtools.
- macOS: Xcode CLI
- Linux: r-base-dev (debian)

I would also add a note that the estimation routines for copulas are written in _C++_ within the JOSS paper. At present, the use of _C++_ isn't stated as it only mentions S4 and R.

#### Contributing

There isn't a Contributor Code of Conduct, contributor guidelines, or how to report a problem list. Perhaps this could be added to the end of the README?

### Vignettes

#### Reproducibility

When using the `r*()` functions, sometimes a code chunk has a `set.seed()` component. On other chunks, this is missing. Thus, there may unintentionally be a hidden state that is present. Moreover, in one vignette, we have the advocation for a non-standard `set_seed()` function (note: the `_` instead of `.`).

https://lrnv.github.io/cort/articles/vignette02_cort_clayton.html#dataset-1

At this point, it would be better to explicitly set a dependency on `R >= 3.6`

#### Styling

Depending on the Vignette, the _R_ code shown either lacks space or has a lot of space.

For example:

- Lacking white space: https://lrnv.github.io/cort/articles/vignette03_ecbkm.html#current-implementation-1
- Spaced out: https://lrnv.github.io/cort/articles/vignette04_bootstrap_varying_m.html#fitting-function-1

Consistency here would be appreciated. Consider running `styler::style_dir("vignettes")`.

### Code

#### API

##### snake_case vs. camelCase

Functions within the package switch back and forth between using snake_case and CamelCase. It would be ideal if there could be some consistency.

https://lrnv.github.io/cort/reference/index.html

Moreover, could the reference portion of the API on the `pkgdown` website be split into subcategories? e.g. data/cort/copulas.

Minor note: `Cort()` uses `N=` whereas elsewhere all function parameters are lower-case.

#### Documentation

There is a documentation article for each function. However, I think some of the documentation entries should receive additional insight into the implementation, improved explanation of what is happening in the example, and perhaps a periodic change in which data set is used for the computation. (~4 ship with the package but only `LifeCycleSavings` from `datasets` is used.)

For instance, in the the titular function `Cort()` for the `cort` package, there is sparse details and no indication of what is contained in a `Cort` object:

https://lrnv.github.io/cort/reference/Cort-Class.html

```r
R> cop <- Cort(LifeCycleSavings[,1:3])
#Splitting...
#
#     1 leaves to split...
#     5 leaves to split...
#     10 leaves to split...
#     1 leaves to split...
#Enforcing constraints...
#Done !
R> str(cop)
#Formal class 'Cort' [package "cort"] with 11 slots
#  ..@ p_value_for_dim_red: num 0.75
#  ..@ number_max_dim     : int 3
#  ..@ min_node_size      : num 1
#  ..@ verbose_lvl        : num 1
#  ..@ vols               : num [1:88] 0.05559 0.16964 0.04987 0.06598 0.00195 ...
#  ..@ f                  : num [1:88] 0 0.02 0.02 0.02 0.02 0 0 0 0.02 0 ...
#  ..@ p                  : num [1:88] 1.95e-09 5.75e-02 2.66e-02 7.37e-03 9.78e-04 ...
#  ..@ a                  : num [1:88, 1:3] 0 0.227 0 0 0 ...
#  ..@ b                  : num [1:88, 1:3] 0.227 1 0.227 0.227 0.227 ...
#  ..@ data               : num [1:50, 1:3] 0.627 0.686 0.824 0.235 0.804 ...
#  ..@ dim                : int 3
```

In contrast, the S4 documentation for `Matrix()` in the `Matrix` package looks like: https://stat.ethz.ch/R-manual/R-devel/library/Matrix/html/Matrix-class.html

In another documentation entry, I think a numerical tolerance issue is present due to using `==` instead of `all.equal()`:
https://lrnv.github.io/cort/reference/quad_prod-methods.html

```r
R> quad_prod(cop,cop)
# [1] 10.81164
R> quad_norm(cop)
# [1] 10.81164
R> quad_norm(cop) == quad_prod(cop,cop)
# [1] FALSE
```

#### Simulated Data

For the simulated data sets, there isn't a simulation script associated with the help file:

- https://lrnv.github.io/cort/reference/clayton_data.html
- https://lrnv.github.io/cort/reference/funcdep_data.html
- https://lrnv.github.io/cort/reference/impossible_data.html
- https://lrnv.github.io/cort/reference/recoveryourself_data.html

Though, the scripts can be found in:

https://github.com/lrnv/cort/tree/master/data-raw

I think it would be helpful to link to `data-raw/` or directly embed it within the Rd file.

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713839737_
