### 2020-11-12 lrnv/cort#8 @lrnv
@pdebuyl review
Some more comments from @pdebuyl:

Typo, page 2, first line. "efficicent" -> "efficient"

Typo, page 2, acknowledgments. "meaningfull" -> "meaningful".

Grammar, page 2, second paragraph. "by statistician" -> "by statisticians" (if you pick singular instead, then conjugation of "need" should be changed as well).

Affiliation: please add "France" to the first affiliation and spell out the second affiliation (+ add city and country).

For specific issues such as the API naming, you can discuss them in cort issues and report the conclusion here. For changes to the paper, I prefer if they are done in this thread.

_Originally posted by @pdebuyl in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-714702564_
