### 2020-11-12 lrnv/cort#9 @lrnv
Readme fixes
### README

#### Installation

Please indicate on the README that installation from GitHub will require the system to have a compiler.

- Windows:  Rtools.
- macOS: Xcode CLI
- Linux: r-base-dev (debian)

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713839737_

Some of the issues that I came across were already mentioned by @coatless above, most importantly, the lack of community guidelines, instructions for potential contributors, for issue reporting or other support. So, I will try not to repeat what has already been said.

_Originally posted by @agisga in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713910360_
