### 2020-11-12 lrnv/cort#10 @lrnv
Vignettes fixes
### Vignettes

#### Reproducibility

When using the `r*()` functions, sometimes a code chunk has a `set.seed()` component. On other chunks, this is missing. Thus, there may unintentionally be a hidden state that is present. Moreover, in one vignette, we have the advocation for a non-standard `set_seed()` function (note: the `_` instead of `.`).

https://lrnv.github.io/cort/articles/vignette02_cort_clayton.html#dataset-1

At this point, it would be better to explicitly set a dependency on `R >= 3.6`

#### Styling

Depending on the Vignette, the _R_ code shown either lacks space or has a lot of space.

For example:

- Lacking white space: https://lrnv.github.io/cort/articles/vignette03_ecbkm.html#current-implementation-1
- Spaced out: https://lrnv.github.io/cort/articles/vignette04_bootstrap_varying_m.html#fitting-function-1

Consistency here would be appreciated. Consider running `styler::style_dir("vignettes")`.

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713839737_
