### 2020-11-12 lrnv/cort#11 @lrnv
API fixes
#### API

##### snake_case vs. camelCase

Functions within the package switch back and forth between using snake_case and CamelCase. It would be ideal if there could be some consistency.

https://lrnv.github.io/cort/reference/index.html

Moreover, could the reference portion of the API on the `pkgdown` website be split into subcategories? e.g. data/cort/copulas.

Minor note: `Cort()` uses `N=` whereas elsewhere all function parameters are lower-case.

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713839737_
