### 2020-11-12 lrnv/cort#13 @lrnv
Simulated dataset fixes
#### Simulated Data

For the simulated data sets, there isn't a simulation script associated with the help file:

- https://lrnv.github.io/cort/reference/clayton_data.html
- https://lrnv.github.io/cort/reference/funcdep_data.html
- https://lrnv.github.io/cort/reference/impossible_data.html
- https://lrnv.github.io/cort/reference/recoveryourself_data.html

Though, the scripts can be found in:

https://github.com/lrnv/cort/tree/master/data-raw

I think it would be helpful to link to `data-raw/` or directly embed it within the Rd file.

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713839737_
