### 2020-11-12 lrnv/cort#6 @lrnv
Splitted into smaller issues
