### 2020-11-12 lrnv/cort#15 @lrnv
Paper fixes
I would also add a note that the estimation routines for copulas are written in _C++_ within the JOSS paper. At present, the use of _C++_ isn't stated as it only mentions S4 and R.

**State of the field**

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

I'm not very familiar with estimation of non-parametric copulas and so this may not be relevant, but I don't really see a comparison to other R or Python packages. A simple web search shows that other copula estimation packages exist. While the `cort` paper mentions publications by Li et al. and Nagler et al. as alternative approaches, it is not clear whether these papers correspond to some commonly-used software packages or are mainly theoretical/methodological papers.

**Quality of writing**

There are some minor grammatical and spelling mistakes in the paper. So, a proofreading would be beneficial. Here are some of the mistakes I noticed:

- Page 1, third paragraph:
    > There also exists several tree-structured piecewise constant density estimators,

    "exists" should be "exist"
     > The new models that are implemented in this package tries to solve these issues.

     "tries" should be 'try"

- Page 2:

     * "pakcage" should be "package"

     * "was design" should be "was designed"

- Page 2:

     > "Examples datasets are included in the package, and the many vignettes gives examples of usecases."

     "Examples" should be "Example", "gives" should be "give", "usecases" should be "use cases".

_Originally posted by @agisga in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-713910360_

Typo, page 2, first line. "efficicent" -> "efficient"

Typo, page 2, acknowledgments. "meaningfull" -> "meaningful".

Grammar, page 2, second paragraph. "by statistician" -> "by statisticians" (if you pick singular instead, then conjugation of "need" should be changed as well).

Affiliation: please add "France" to the first affiliation and spell out the second affiliation (+ add city and country).

For specific issues such as the API naming, you can discuss them in cort issues and report the conclusion here. For changes to the paper, I prefer if they are done in this thread.

_Originally posted by @pdebuyl in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-714702564_
