### 2020-11-12 lrnv/cort#7 @lrnv
Splitted into smaller issues
