### 2020-11-12 lrnv/cort#8 @lrnv
Did get merged with the paper issue #15
