### 2020-11-13 70a6d16
Adding info to the readme about the need of a compiler for developement version
