+++ README.md
@@ -29,6 +29,12 @@ The upstream development version can also be installed with :
 devtools::install_github("lrnv/cort")
 ```
 
+Note that the installation from github will require the system to have a compiler: 
+
+- Windows: Rtools
+- macOS: Xcode CLI
+- Linux: r-base-dev (debian)
+
 
 The vignettes are quite expressive. They give a clear overview of what can be done with this package, how it is coded and why it is useful. Please read them for more details. 
 
