### 2020-11-13 903ee3b
Adding comunities guidelines, intruction for potnetial contributors, issue reporting or other support.
