+++ README.md
@@ -38,6 +38,14 @@ Note that the installation from github will require the system to have a compile
 
 The vignettes are quite expressive. They give a clear overview of what can be done with this package, how it is coded and why it is useful. Please read them for more details. 
 
+## How to report bugs and get support
+
+To report a bug, feel free to open an issue on the github repository. Support can also be provided thorugh the same chanel if you need it.
+
+## How to contribute
+
+Every contribution is welcome, on the form of pull requests on the github repository. For large modifications, please open an issue for discussions firsts. 
+
 
 ## References
 
