+++ tests/testthat/test-CortForest.R
@@ -2,7 +2,6 @@ context("Testing the class CortForest...")
 
 quiet(library(cort))
 quiet(suppressWarnings(library(purrr)))
-furrr::future_options(seed=TRUE)
 model = CortForest(LifeCycleSavings[,1:2],verbose_lvl = 0,n_trees=3)
 
 quiet(show(model))
