### 2020-11-13 4d841ce
Remove the weird set_seed part. Now the set.seed mchanisme is used in all vignettes the same way.
