+++ vignettes/vignette02_cort_clayton.Rmd
@@ -40,23 +40,10 @@ rClayton <- function(n,dim,alpha){
 ```
 
 
-For reproducibility reasons, we set the random number generator. This vignette has been compiled on a version of R > 3.6, so the fix of the random number generator added in this version is used. To reproduce results from an earlier version, just use `sample.kind = "Rounding"`.
-
-```{r setseed}
-set_seed <- function(seed){
-  if(as.numeric(version$minor)<6){
-    # the way of specifying the random number generation changed. 
-    set.seed(seed,kind = "Mersenne-Twister",normal.kind = "Inversion")
-  } else {
-    set.seed(seed,kind = "Mersenne-Twister",normal.kind = "Inversion",sample.kind = "Rejection")
-  }
-}
-```
-
 The following code simulates a dataset and then visualise it : 
 
 ```{r dataset} 
-set_seed(12)
+set.seed(12)
 n = 200 # taken small to reduce runtime of the vignette.
 d = 4
 n_trees = 5 # taken small to reduce runtime of the vignette.
