### 2020-11-13 e9c5c35
Improvements in spacing consistency of the code in the vignettes
