+++ vignettes/vignette01_ecb.Rmd
@@ -73,6 +73,7 @@ You need to provide a dataset, defining $\hat{\mu}$, to construct the model. For
 set.seed(1)
 data("LifeCycleSavings")
 pseudo_data <- (apply(LifeCycleSavings,2,rank,ties.method="max")/(nrow(LifeCycleSavings)+1))
+
 pairs(pseudo_data,lower.panel=NULL)
 ```
 
@@ -89,6 +90,7 @@ For the moment, only some methods exist for this copula. We can calculate it's v
 
 ```{r, fig.cap = "Pairs-plot of original peusdo-observation from the data (red) with simulated pseudo_observation (black)"}
 simu <- rCopula(n = 1000,copula = cop)
+
 pairs(rbind(simu,pseudo_data),
       col=c(rep("black",nrow(simu)),rep("red",nrow(pseudo_data))),
       gap=0,
