+++ R/data.R
@@ -8,6 +8,15 @@
 #'
 #' @format A matrix with 500 rows and 3 columns
 #'
+#' The example section below gives the code to re-generate this data if needed.
+#'
+#' @examples
+#' set.seed(seed = 12,kind = "Mersenne-Twister",normal.kind = "Inversion")
+#' x = matrix(runif(1500),500,3)
+#' x[,2] = sin(2*pi*x[,1])-x[,2]/pi
+#' x[,3] = (x[,3]*(x[,1]<1/4)/2 - sin(pi**(x[,1]))*(x[,1]>1/4))*(1+x[,3]/(pi^2))
+#' funcdep_data = apply(x,2,function(x){return(rank(x,ties.method = "max"))})/(501)
+#'
 #' @references
 #' \insertRef{laverny2020}{cort}
 "funcdep_data"
@@ -21,6 +30,23 @@
 #' This dataset is studied in O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020).
 #'
 #' @format A matrix with 200 rows and 2 columns
+#'
+#' The example section below gives the code to re-generate this data if needed.
+#'
+#' @examples
+#' set.seed(seed = 12, kind = "Mersenne-Twister", normal.kind = "Inversion")
+#' x = matrix(runif(400),200,2)
+#' x = t(apply(x, 1,function(u){
+#'   if(u[1]< 1/3){
+#'     u[2] = 1/2 + u[2]/2
+#'   } else{ if(u[1]<2/3){
+#'     u[2] = u[2]/2
+#'   } else {
+#'     u[2] = 1/2 + u[2]/2
+#'   }}
+#'   return(u)
+#' }))
+#' impossible_data = apply(x,2,function(x){return(rank(x,ties.method = "max"))})/(201)
 #'
 #' @references
 #' \insertRef{laverny2020}{cort}
@@ -38,6 +64,24 @@
 #'
 #' @format A matrix with 500 rows and 2 columns
 #'
+#' The example section below gives the code to re-generate this data if needed.
+#'
+#' @examples
+#' set.seed(seed = 12, kind = "Mersenne-Twister", normal.kind = "Inversion")
+#' x = matrix(runif(1000),500,2)
+#' recoveryourself_data = t(apply(x, 1,function(u){
+#'   if(u[1]< 1/4){
+#'     u[2] = 3/4 + u[2]/4
+#'   } else{ if(u[1]<1/2){
+#'     u[2] = 1/2 + u[2]/4
+#'   } else { if(u[1]<3/4){
+#'     u[2] = u[2]/4
+#'   } else {
+#'     u[2] = 1/4 + u[2]/4
+#'   }}}
+#'   return(u)
+#' }))
+#'
 #' @references
 #' \insertRef{laverny2020}{cort}
 "recoveryourself_data"
@@ -52,6 +96,22 @@
 #' This dataset is studied in O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020).
 #'
 #' @format A matrix with 200 rows and 4 columns
+#'
+#' The example section below gives the code to re-generate this data if needed.
+#'
+#' @examples
+#' psi <- function(t,alpha) (1 + sign(alpha)*t) ^ (-1/alpha) # generator
+#' rClayton <- function(n,dim,alpha){
+#'   val <- matrix(runif(n * dim), nrow = n)
+#'   gam <- rgamma(n, shape = 1/alpha, rate = 1)
+#'   gam <- matrix(gam, nrow = n, ncol = dim)
+#'   psi(- log(val) / gam,alpha)
+#' }
[…]
