### 2020-11-13 e20913d
Fixing the reference page of the pkgdown website with categories.
