### 2020-11-13 lrnv/cort#11 @lrnv
Fixing the reference index, you can now see that the CamelCase and snake_case switches are to seperate constructors and standard (r/d/p/v)-functions from methods and the rest. Take a look at the new reference, you will see that the common uses of the two standards makes sense now that it's ordered.

For the last minor note : N is the number of bootstraps, and is intensionally given in uppercase as it is common in `R`, see e.g `boot::boot`, and many others.
