### 2020-11-13 bd1b58b
First part of documentaiton improvements.
