+++ R/generics.R
@@ -58,7 +58,7 @@ setGeneric("vCopula", function(u, v, copula, ...) {
 #' @rdname dCopula-methods
 #'
 #' @examples
-#' cop <- cbCopula(LifeCycleSavings,m = 5)
+#' cop <- cbCopula(Cort::funcdep_data[1:10,1:2], m = 5)
 #' dCopula(rep(0,5),cop)
 #' dCopula(rep(0.5,5),cop)
 #' dCopula(rep(1,5),cop)
@@ -68,7 +68,7 @@ setGeneric("dCopula", function(u, copula, ...) {
   standardGeneric("dCopula")
 })
 
-#' Copula density
+#' Copula cdf
 #'
 #' This function returns the value of the copula itself on given points.
 #'
@@ -83,7 +83,7 @@ setGeneric("dCopula", function(u, copula, ...) {
 #' @rdname pCopula-methods
 #'
 #' @examples
-#' cop <- cbCopula(LifeCycleSavings,m = 5)
+#' cop <- cbCopula(Cort::recoveryourself_data,m = 5)
 #' pCopula(rep(0,5),cop) == 0
 #' pCopula(rep(0.5,5),cop)
 #' pCopula(rep(1,5),cop) == 1
@@ -128,7 +128,7 @@ setMethod("vCopula", signature = c(u = "matrix", v = "matrix"),
           })
 
 
-#' Copula random variables simulation
+#' Copula random generation
 #'
 #' This function simulate random variables from a copula.
 #'
@@ -143,7 +143,7 @@ setMethod("vCopula", signature = c(u = "matrix", v = "matrix"),
 #' @rdname rCopula-methods
 #'
 #' @examples
-#' cop <- cbCopula(LifeCycleSavings,m = 5)
+#' cop <- cbCopula(Cort::clayton_data,m = 5)
 #' xx <- rCopula(1000,cop)
 #'
 setGeneric("rCopula", function(n, copula, ...) standardGeneric("rCopula"))
@@ -180,7 +180,7 @@ setGeneric("biv_rho", function(copula) standardGeneric("biv_rho"))
 #' @rdname biv_tau-methods
 #'
 #' @examples
-#' cop <- Cort(LifeCycleSavings[,1:3])
+#' cop <- Cort(Cort::funcdep_data[,1:3])
 #' biv_tau(cop)
 #'
 setGeneric("biv_tau", function(copula) standardGeneric("biv_tau"))
@@ -188,8 +188,9 @@ setGeneric("biv_tau", function(copula) standardGeneric("biv_tau"))
 
 
 
-#' Loss of the model
+#' Loss of a copula estimation (if the model has one)
 #'
+#' Currently only implemented for Cort models.
 #' Compute the loss of the model
 #'
 #'
@@ -201,14 +202,15 @@ setGeneric("biv_tau", function(copula) standardGeneric("biv_tau"))
 #' @rdname loss-methods
 #'
 #' @examples
-#' cop <- Cort(LifeCycleSavings[,1:3])
+#' cop <- Cort(Cort::recoveryourself_data)
 #' loss(cop)
 #'
 setGeneric("loss", function(object) standardGeneric("loss"))
 
 
-#' Constraint influence of the model
+#' Constraint influence of the model (if it has one)
 #'
+#' Currently only implemented for Cort models.
 #' Compute the constraint influence of the model
 #'
 #'
@@ -220,13 +222,14 @@ setGeneric("loss", function(object) standardGeneric("loss"))
 #' @rdname constraint_infl-methods
 #'
 #' @examples
-#' cop <- Cort(LifeCycleSavings[,1:3])
+#' cop <- Cort(Cort::funcdep_data[,1:3])
[…]
