+++ R/CortForest.R
@@ -27,7 +27,10 @@
 
 #' CortForest class
 #'
-#' This class implements the bagging of CORT models, with an oob error minimisation in the weights. See O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020) for the details of this density estimation procedure.
+#' This class implements the bagging of CORT models, with an out-of-bag error minimisation in the weights.
+#'
+#'
+#' See O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020) for the details of this density estimation procedure.
 #'
 #'
 #' @param x The data, must be provided as a matrix with each row as an observation.
@@ -42,7 +45,7 @@
 #' @param oob_weighting boolean (default : TRUE) option to weight the trees with an oob criterion (otherwise they are equally weighted)
 #'
 #' @name CortForest-Class
-#' @title Bagged Cort estimates
+#' @title Bagged Cort copulas
 #' @rdname CortForest-Class
 #'
 #' @return a CortForest object that can be fitted easily to produce a copula estimate.
