+++ man/CortForest-Class.Rd
@@ -3,7 +3,7 @@
 \name{CortForest-Class}
 \alias{CortForest-Class}
 \alias{CortForest}
-\title{Bagged Cort estimates}
+\title{Bagged Cort copulas}
 \usage{
 CortForest(
   x,
@@ -46,7 +46,9 @@ a CortForest object that can be fitted easily to produce a copula estimate.
 CortForest class
 }
 \details{
-This class implements the bagging of CORT models, with an oob error minimisation in the weights. See O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020) for the details of this density estimation procedure.
+This class implements the bagging of CORT models, with an out-of-bag error minimisation in the weights.
+
+See O. Laverny, V. Maume-Deschamps, E. Masiello and D. Rullière (2020) for the details of this density estimation procedure.
 }
 \examples{
 (CortForest(LifeCycleSavings[,1:3],number_max_dim=2,n_trees=2))
