+++ man/dCopula-methods.Rd
@@ -24,7 +24,7 @@ dCopula(u, copula, ...)
 \item{...}{other parameter to be passed to methods for this generic.}
 }
 \value{
-the density of the copula on each observation
+The density of the copula on each observation
 }
 \description{
 This function returns the density of a given copula on given observations.
