+++ man/rCopula-methods.Rd
@@ -30,10 +30,10 @@ rCopula(n, copula, ...)
 \item{...}{other parameter to be passed to methods for this generic.}
 }
 \value{
-the density of the copula on each observation
+A matrix with \code{n} rows, each representing a random vetor generated from the provided copula.
 }
 \description{
-This function simulate random variables from a copula.
+Random number generation following the given copula. This function performs the simulation of random vectors following the copula.
 }
 \section{Functions}{
 \itemize{
