### 2020-11-16 f42c1c7
Updating tic usage. Hopefully this should work.
