### 2020-11-16 4c7d9b4
First part of paper fixes
