### 2020-11-16 #2653 @lrnv
@pdebuyl , I am at a stage were everything that needed to be fixed EXCEPT the paper is fixed. Since the paper is not part of the CRAN submission, assuming that the reviewers have nothing more to say about the code itself, I'll update the version number and push to cran.

We can then discuss the paper while CRAN do it's job, and finish the publication to zenodo / Joss when CRAN has accepted.

Would you agree with this plan ? Did I forgot something ?
