+++ NEWS.md
@@ -1,4 +1,13 @@
-# cort (development version)
+# cort 0.3.2
+
+## Minor improvements and fixes
+
+* Improved documentation
+* Fixed buggy automated tests on CortForest
+* Added simulation script to the documentation of the datasets
+* Cleared up pkgdown reference page
+* Cleaned up some vignettes
+* README fixes.
 
 # cort 0.3.1
 
