### 2020-11-16 #2653 @lrnv
@coatless , @agisga
I wanted to thank again both of you for the time you took reviewing this package. I made modifications regarding all your remarks and suggestions, improving in my opinion the quality of the final code and documentation.

I'm now ready to submit an update to CRAN. If you want to see what has changed, feel free to open the corresponding issues. If you have more remarks regarding the code, please tell me. Otherwise, please tell me also, so I can push to CRAN.
