### 2020-11-16 #2653 @pdebuyl
@lrnv the plan to get validation of the code first is ok for me.

@coatless @agisga can you proceed first to the code checks so that @lrnv can submit the package to CRAN?
