### 2020-11-19 #2653 @lrnv
@agisga Thanks a lot ! I will address all this, of course.

It's funny that you get errors from furrr, as none of the testing plateforms I used tells that -- maybe my tests are not covering what you did, so If you remember what you were doing we could add it as a test ? Also, do you remember your version of R and furrr before updating ? I will try to understand when the change to furrr happend and set a dependency accordingly.
