### 2020-11-19 #2653 @agisga
@lrnv I don't think this is a serious issue, since everything works with the current version of furrr from CRAN. I think the error was related to `furrr::furrr_options`. According to [this post](https://www.tidyverse.org/blog/2020/10/furrr-0-2-0/), the name `furrr_options` was introduced in furrr v0.2.0 just a month ago (current version on CRAN is 0.2.1). I must have had an older version of furrr installed (not sure which exactly). I'm not sure whether, if cort had a requirement on furrr versioning, it would have been updated automatically when I built/installed cort (but I would expect this to be the case; that's why I suggested to provide version requirements for the dependencies in the DESCRIPTION, at least for furrr).
