### 2020-11-19 lrnv/cort#26 @lrnv
@agisga 2nd round
@lrnv Generally the code changes look good. Here are a few comments based on my previous feedback and the new changes:

- Vignette 4 still has syntactically wrong and redundant two lines of code here: https://github.com/lrnv/cort/blob/dd6d5abeabd10ed3cab1fbdbceaac0725081d912/vignettes/vignette04_bootstrap_varying_m.Rmd#L55
- In issue https://github.com/lrnv/cort/issues/11 you say that "the CamelCase and snake_case switches are to seperate constructors and standard (r/d/p/v)-functions from methods and the rest. Take a look at the new reference, you will see that the common uses of the two standards makes sense now that it's ordered." It would be good to mention these or other coding/syntax/style standard (if any) that you follow under the contribution guidelines in the README (or a separate file if it's too much information for the README).
- The new version of the `cort` package works flawlessly on my system. However, the first time I tried it, I was getting error messages, which went away after I upgraded `furrr` to the newest version (from CRAN). I'm not sure what version of `furrr` I had prior to upgrading. Maybe it would be good to provide version requirements for `furrr` and all other dependencies in the [`DESCRIPTION`](https://github.com/lrnv/cort/blob/master/DESCRIPTION) (currently, `testthat` has a version requirement stated, but other dependencies do not)?

I also had a very quick look at the new changes to the paper:

- The proper citation info for the `copula` R package is given at https://cran.r-project.org/web/packages/copula/citation.html I don't think it's appropriate to just give a link, when there are several peer-reviewed JSS publications.

_Originally posted by @agisga in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-729958270_
