### 2020-11-19 ef1e68f
Fixing the copula citation
