+++ paper.bib
@@ -63,3 +63,44 @@
   journal={arXiv preprint arXiv:2005.02912},
   year={2020}
 }
+
+  @Manual{cop1,
+    title = {copula: Multivariate Dependence with Copulas},
+    author = {Marius Hofert and Ivan Kojadinovic and Martin Maechler
+      and Jun Yan},
+    year = {2020},
+    note = {R package version 1.0-0},
+    url = {https://CRAN.R-project.org/package=copula},
+  }
+  @Article{cop2,
+    title = {Enjoy the Joy of Copulas: With a Package {copula}},
+    author = {{Jun Yan}},
+    journal = {Journal of Statistical Software},
+    year = {2007},
+    volume = {21},
+    number = {4},
+    pages = {1--21},
+    url = {http://www.jstatsoft.org/v21/i04/},
+  }
+  @Article{cop3,
+    title = {Modeling Multivariate Distributions with Continuous
+      Margins Using the {copula} {R} Package},
+    author = {{Ivan Kojadinovic} and {Jun Yan}},
+    journal = {Journal of Statistical Software},
+    year = {2010},
+    volume = {34},
+    number = {9},
+    pages = {1--20},
+    url = {http://www.jstatsoft.org/v34/i09/},
+  }
+  @Article{cop4,
+    title = {Nested Archimedean Copulas Meet {R}: The {nacopula}
+      Package},
+    author = {{Marius Hofert} and {Martin M\"achler}},
+    journal = {Journal of Statistical Software},
+    year = {2011},
+    volume = {39},
+    number = {9},
+    pages = {1--20},
+    url = {http://www.jstatsoft.org/v39/i09/},
+  }
