### 2020-11-24 #2653 @pdebuyl
@agisga @coatless the non-paper bullet points missing from the checklist concern installation and community guidelines. Are these ok? If so, @lrnv can proceed to CRAN submission.

@lrnv I had to install a binary package for nloptr in addition to `install.packages("cort")` (on debian "r-cran-nloptr"). I don't know if it is expected of R users to guess that some components require a separate installation.
