### 2020-11-24 #2653 @lrnv
@pdebuyl Well, concerning `nloptr`, i dont know. They only say on their readme [there](https://cran.r-project.org/web/packages/nloptr/readme/README.html) that caution must be taken on windows, and nothing for debian. Although a simple google search shows that you are not the only one with this kind of issues. It should work as expected: installation of the R package *should* download `nlopt`, the binary dependency. But sometimes, epsecially in debian which is quite conservative, this does not work that way. Usually no problems on Arch.

Anyway, this is nloptr problem, i guess that "users are expected to guess that some components require a separate installation", sadly. This has been a known issue since a while as far as what i can read, e.g (there)[https://github.com/jyypma/nloptr/issues/40].
