### 2020-11-25 #2653 @agisga
@pdebuyl Yes, I'm okay with all of the non-paper bullet points now.
