### 2020-11-25 #2653 @coatless
@lrnv thanks for making the changes. I just have two notes:

Paper-wise, there is a missing note that the underlying non-parametric algorithms are written in _C++_. It would be great to link out to the `RcppArmadillo`/`Rcpp` paper here.

Regarding code, the documentation governing the object slots is still a bit sparse. It would be good to understand what is being returned within the `Cort()` object explicitly.

That is, what do the following slot entries correspond to?

```
p_value_for_dim_red
number_max_dim
min_node_size
verbose_lvl
vols
f
p
a
b
data
dim
```
