### 2020-11-26 lrnv/cort#28 @lrnv
@coatless comments: Paper
@coatless wrote:

Paper-wise, there is a missing note that the underlying non-parametric algorithms are written in _C++_. It would be great to link out to the `RcppArmadillo`/`Rcpp` paper here.

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-733910858_
