### 2020-11-26 lrnv/cort#29 @lrnv
@coatless comments: Docs
@coatless wrote

Regarding code, the documentation governing the object slots is still a bit sparse. It would be good to understand what is being returned within the `Cort()` object explicitly.

That is, what do the following slot entries correspond to?

```
p_value_for_dim_red
number_max_dim
min_node_size
verbose_lvl
vols
f
p
a
b
data
dim
```

_Originally posted by @coatless in https://github.com/openjournals/joss-reviews/issues/2653#issuecomment-733910858_
