### 2020-11-26 7fab094
Adding documentation for slots of S4 objects. Only Cort and CortForest needed some, the other objects were alrady fully covered (parameters of the constructor == slots).
