### 2020-11-26 4e38419
Fix #28. We added citation for Rcpp, but also for Future, and we said that the computations were done mainly in C++.
