+++ paper.md
@@ -33,7 +33,7 @@ We note that a lot of tools are avaliable in R for copula modelling  through the
 
 The Copula recursive tree (cort) designed by [@laverny2020dependence] is a flexible, consistent, piecewise linear estimator for a copula [@sklar1959fonctions], leveraging the patchwork copula formalization [@durante2015convergence] and a specific piecewise constant density estimator, the density estimation tree [@ram2011density]. While the patchwork structure imposes the grid, this estimator is data-driven and constructs the grid recursively from the data, minimizing a chosen distance on the copula space. Furthermore, while the addition of the copula constraints makes the available solutions for density estimation unusable, our estimator is only concerned with dependence and guarantees the uniformity of margins. The R package `cort` provides a useful implementation of this model and several potential refinements, allowing for fast computations of cort trees, and parallel computations of cort forests.
 
-The main feature implemented in the package is the cort algorithm, a non-parametric, piecewise constant, copula density estimator. The implementation is recursive and hence quite efficient. The `cort` package is a statistical package that allows to estimate several non-parametric copula models in R. Although the state of the art `copula` package has functions to estimate the empirical copula, we provide a structured set of S4 classes that allows estimation of empirical copulas, checkerboard copulas, cort copula and bagging of all of these. In the cort forest, the implementation allows the code to run in parallel, to fasten the computations.
+The main feature implemented in the package is the cort algorithm, a non-parametric, piecewise constant, copula density estimator. The implementation is recursive and hence quite efficient. The `cort` package is a statistical package that allows to estimate several non-parametric copula models in R. Although the state of the art `copula` package has functions to estimate the empirical copula, we provide a structured set of S4 classes that allows estimation of empirical copulas, checkerboard copulas, cort copula and bagging of all of these. A specific class exists for bagging cort models, which implementation runs in parallel, to fasten the computations, using the `future` package [@future]. Most of the underlying machinery and computations are written in `C++`, through the `Rcpp` [@rccp1,@rccp2,@rcpp3] package.
 
 The `cort` package was designed to be used by statisticians who need a non-parametric view of the dependence structure of a given dataset. It features a rich and extensive API to call the statistic fitting procedures, plotting functions and tools to assess the quality of the fits, while complying with the R standards. Examples datasets are included in the package, and the many vignettes give examples of use cases.
 
