+++ paper.bib
@@ -104,3 +104,47 @@
     pages = {1--20},
     url = {http://www.jstatsoft.org/v39/i09/},
   }
+  @Misc{future,
+    author = {Henrik Bengtsson},
+    title = {A Unifying Framework for Parallel and Distributed
+      Processing in R using Futures},
+    year = {2020},
+    month = {aug},
+    eprint = {2008.00553},
+    archiveprefix = {arXiv},
+    primaryclass = {cs.DC},
+    url = {https://arxiv.org/abs/2008.00553},
+  }
+  @Article{rcpp1,
+    title = {{Rcpp}: Seamless {R} and {C++} Integration},
+    author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
+    journal = {Journal of Statistical Software},
+    year = {2011},
+    volume = {40},
+    number = {8},
+    pages = {1--18},
+    url = {http://www.jstatsoft.org/v40/i08/},
+    doi = {10.18637/jss.v040.i08},
+  }
+  @Book{rcpp2,
+    title = {Seamless {R} and {C++} Integration with {Rcpp}},
+    author = {Dirk Eddelbuettel},
+    publisher = {Springer},
+    address = {New York},
+    year = {2013},
+    note = {ISBN 978-1-4614-6867-7},
+    doi = {10.1007/978-1-4614-6868-4},
+  }
+  @Article{rcpp3,
+    title = {{Extending extit{R} with extit{C++}: A Brief Introduction
+      to extit{Rcpp}}},
+    author = {Dirk Eddelbuettel and James Joseph Balamuta},
+    journal = {PeerJ Preprints},
+    year = {2017},
+    month = {aug},
+    volume = {5},
+    pages = {e3188v1},
+    issn = {2167-9843},
+    url = {https://doi.org/10.7287/peerj.preprints.3188v1},
+    doi = {10.7287/peerj.preprints.3188v1},
+  }
