### 2020-11-26 #2653 @coatless
@lrnv could you use the following Rcpp citations?

```tex
@Article{,
  title = {{Rcpp}: Seamless {R} and {C++} Integration},
  author = {Dirk Eddelbuettel and Romain Fran\c{c}ois},
  journal = {Journal of Statistical Software},
  year = {2011},
  volume = {40},
  number = {8},
  pages = {1--18},
  url = {http://www.jstatsoft.org/v40/i08/},
  doi = {10.18637/jss.v040.i08},
}

@Article{,
  title = {RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
  author = {Dirk Eddelbuettel and Conrad Sanderson},
  journal = {Computational	Statistics and Data Analysis},
  year = {2014},
  volume = {71},
  month = {March},
  pages = {1054--1063},
  url = {http://dx.doi.org/10.1016/j.csda.2013.02.005},
}
```
