+++ paper.bib
@@ -148,3 +148,13 @@
     url = {https://doi.org/10.7287/peerj.preprints.3188v1},
     doi = {10.7287/peerj.preprints.3188v1},
   }
+  @Article{rccp4,
+  title = {RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
+  author = {Dirk Eddelbuettel and Conrad Sanderson},
+  journal = {Computational	Statistics and Data Analysis},
+  year = {2014},
+  volume = {71},
+  month = {March},
+  pages = {1054--1063},
+  url = {http://dx.doi.org/10.1016/j.csda.2013.02.005},
+}
