### 2020-11-26 9b5f2d4
tryig to have whedon generate all citations.
