+++ paper.bib
@@ -148,7 +148,7 @@
     url = {https://doi.org/10.7287/peerj.preprints.3188v1},
     doi = {10.7287/peerj.preprints.3188v1},
   }
-  @Article{rccp4,
+  @Article{rcpp4,
   title = {RcppArmadillo: Accelerating R with high-performance C++ linear algebra},
   author = {Dirk Eddelbuettel and Conrad Sanderson},
   journal = {Computational	Statistics and Data Analysis},
