### 2020-11-26 #2653 @lrnv
@coatless I fixed the references.

@pdebuyl I'll proceed to cran then if everyone agrees.
