### 2020-11-26 #2653 @coatless
@lrnv minor note on the citation: I think text for `RcppArmadillo` is missing after the `and`:

![Screen Shot 2020-11-26 at 2 28 28 PM](https://user-images.githubusercontent.com/833642/100389713-bf6bbf00-2ff3-11eb-8572-0d824d9d3744.png)

Also, if you're aiming to keep the extending citation, could you tweak it slight to avoid having `exit` appear in the title?

```
@Article{,
  title = {{Extending R with C++: A Brief Introduction to Rcpp}},
  author = {Dirk Eddelbuettel and James Joseph Balamuta},
  journal = {PeerJ Preprints},
  year = {2017},
  month = {aug},
  volume = {5},
  pages = {e3188v1},
  issn = {2167-9843},
  url = {https://doi.org/10.7287/peerj.preprints.3188v1},
  doi = {10.7287/peerj.preprints.3188v1},
}
```

would fix:

![Screen Shot 2020-11-26 at 2 32 10 PM](https://user-images.githubusercontent.com/833642/100389856-30ab7200-2ff4-11eb-87ed-1d507c7fff6b.png)
