### 2020-11-26 #2653 @pdebuyl
@lrnv good, let us know when the CRAN process is complete.
