+++ NEWS.md
@@ -1,3 +1,5 @@
+# cort (development version)
+
 # cort 0.3.2
 
 ## Minor improvements and fixes
