### 2020-12-01 f3506a7
spell checking paper.md
