+++ paper.bib
@@ -127,7 +127,7 @@
     doi = {10.18637/jss.v040.i08},
   }
   @Book{rcpp2,
-    title = {Seamless {R} and {C++} Integration with {Rcpp}},
+    title = {Seamless R and C++ Integration with {Rcpp}},
     author = {Dirk Eddelbuettel},
     publisher = {Springer},
     address = {New York},
@@ -136,7 +136,7 @@
     doi = {10.1007/978-1-4614-6868-4},
   }
   @Article{rcpp3,
-    title = {{Extending extit{R} with extit{C++}: A Brief Introduction
+    title = {{Extending R with C++: A Brief Introduction
       to extit{Rcpp}}},
     author = {Dirk Eddelbuettel and James Joseph Balamuta},
     journal = {PeerJ Preprints},
