### 2020-12-01 #2653 @lrnv
@pdebuyl The version 0.3.2 is now on CRAN ! I also made some spellchecks on the paper.

What would be the next step ?
