### 2020-12-01 #2653 @pdebuyl
Thanks @lrnv

Can you
1. Group the multiple citations using the syntax `[@a;@b]` ?
2. Mention the paper's availability on cran?

We just need the final checks from @agisga (checklist items at the top of the issue + explicit accept of the paper). @coatless approved already. Then, I'll proceed with the editorial part.

You also need to archive it on zenodo (or another doi-granting archival platform such as figshare) and post the doi here.
