### 2020-12-01 #2653 @agisga
@pdebuyl @lrnv My concerns with the paper have been addressed. I think the paper can be accepted in this form.

Here are just a few minor grammar-related suggestions for your consideration:
- "Copulas are distributions functions" ~~~> "Copulas are **distribution** functions".
- "Although the estimation of copulas is a wide-treated subject" ~~> maybe better to say "widely-treated".
- "The Copula recursive tree, a.k.a Cort, " ~~> maybe better "The Copula recursive tree, **or** Cort,".
- "extensive API to call the statistic fitting procedures" ~~> "extensive API to call the **statistical** fitting procedures".
