### 2020-12-01 #2653 @pdebuyl
Thank you @agisga !

@lrnv the zenodo entry has the tag 3.2.1 and states "Version 3.2.1". Was there an unintended change of the version definition?
