+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: cort
 Title: Some Empiric and Nonparametric Copula Models
-Version: 0.3.2.9000
+Version: 0.3.2
 Authors@R: 
     person(given = "Oskar",
            family = "Laverny",
