### 2020-12-02 #2653 @pdebuyl
I take the second archive. The doi points to the specific release, so there is no technical problem here. You can go to the "release" github page and trigger the publishing mechanism for tags that have not been released.
