### 2020-12-02 #2653 @pdebuyl
@lrnv it is ok. I will proceed to the acceptance work (JOSS is well documented: https://joss.readthedocs.io/en/latest/editing.html#after-reviewers-recommend-acceptance ), which should be done by tomorrow.
