+++ paper.bib
@@ -81,6 +81,7 @@
     number = {4},
     pages = {1--21},
     url = {http://www.jstatsoft.org/v21/i04/},
+    doi = {10.18637/jss.v021.i04},
   }
   @Article{cop3,
     title = {Modeling Multivariate Distributions with Continuous
