### 2020-12-03 #2653 @lrnv
I added the missing DOI. Reference checking seems to be a little bugged !
