### 2020-12-03 #2653 @pdebuyl
Thanks. Could you add the arxiv link to the papers missing one, as it is done for Bengtsson 2020 ? Also, can you use the DOI instead of the url for the J Stat Soft papers? DOIs should be the most stable identifiers and are preferred.

Ram & Gray has doi 10.1145/2020408.2020507 can you use it?
