### 2020-12-03 1cc3868
adding urls and dois
