+++ paper.bib
@@ -15,7 +15,8 @@
   author={Ram, Parikshit and Gray, Alexander G},
   booktitle={Proceedings of the 17th ACM SIGKDD international conference on Knowledge discovery and data mining},
   pages={627--635},
-  year={2011}
+  year={2011},
+  doi={10.1145/2020408.2020507},
 }
 
 @article{wu2018density,
@@ -30,10 +31,10 @@
   doi={10.1080/03610926.2017.1285929}
 }
 
-@article{sklar1959fonctions,
-  title={Fonctions de repartition an dimensions et leurs marges},
-  author={Sklar, M},
-  journal={Publ. inst. statist. univ. Paris},
+@article{sklar1959fonction,
+  title={Fonction de r{\'e}partition dont les marges sont donn{\'e}es},
+  author={Sklar, Abe},
+  journal={Inst. Stat. Univ. Paris},
   volume={8},
   pages={229--231},
   year={1959}
@@ -50,18 +51,24 @@
   doi={10.1016/j.jmva.2016.07.003}
 }
 
-@article{li2018panda,
-  title={PANDA: AdaPtive Noisy Data Augmentation for Regularization of Undirected Graphical Models},
-  author={Li, Yinan and Liu, Xiao and Liu, Fang},
-  journal={arXiv preprint arXiv:1810.04851},
-  year={2018}
+@misc{li2018panda,
+  title={PANDA: AdaPtive Noisy Data Augmentation for Regularization of Undirected Graphical Models}, 
+  author={Yinan Li and Xiao Liu and Fang Liu},
+  year={2019},
+  eprint={1810.04851},
+  archivePrefix={arXiv},
+  primaryClass={stat.ML},
+  url={https://arxiv.org/abs/1810.04851}
 }
 
-@article{laverny2020dependence,
-  title={Dependence structure estimation using Copula Recursive Trees},
-  author={Laverny, Oskar and Maume-Deschamps, V{\'e}ronique and Masiello, Esterina and Rulli{\`e}re, Didier},
-  journal={arXiv preprint arXiv:2005.02912},
-  year={2020}
+@misc{laverny2020dependence,
+  title={Dependence structure estimation using Copula Recursive Trees}, 
+  author={Oskar Laverny and Véronique Maume-Deschamps and Esterina Masiello and Didier Rullière},
+  year={2020},
+  eprint={2005.02912},
+  archivePrefix={arXiv},
+  primaryClass={math.ST},
+  url={https://arxiv.org/abs/2005.02912}
 }
 
   @Manual{cop1,
