### 2020-12-03 #2653 @lrnv
I added some urls and doi in the bib file. Sklar's paper is too old to have any of these unfortunately;
