### 2020-12-03 #2653 @pdebuyl
The reference checker is an indicative first pass. It's almost there https://github.com/lrnv/cort/pull/31
