### 2020-12-04 #2653 @pdebuyl
Thanks @coatless @agisga for the review, @lrnv for submitting to JOSS! The editor in chief in rotation will take over from now.
